+++ src/lexer.ts
@@ -25,7 +25,7 @@ const MATCH_TAG_OPEN = /^<(\/?)([a-zA-Z0-9\-:]+)/;       // https://www.w3.org/T
 const MATCH_TAG_CLOSE = /^\/?>/;
 const MATCH_TEXT = /^[^]*?(?=([ \t]*\n|<|$))/;
 const MATCH_TAG_LOOKAHEAD = /^[^]*?(?=<|$)/;
-const MATCH_ATTR_START = /^([^\t\n\f \/>"'=]+)/;         // https://www.w3.org/TR/html/syntax.html#elements-attributes
+const MATCH_ATTR_START = /^([^\t\n\f \/><"'=]+)/;        // https://www.w3.org/TR/html/syntax.html#elements-attributes
 const MATCH_ATTR_SINGLE = /^='([^']*?)(')/;
 const MATCH_ATTR_DOUBLE = /^="([^"]*?)(")/;
 const MATCH_ATTR_UNQUOTED = /^=([a-zA-Z0-9]+)/;